requirements.txt
Flask
flask-sqlalchemy
flask-login
textblob
Pillow
pytesseract
pandas
scikit-learn
gunicorn